Code cleanup

+++ src/test/java/org/sonar/plugins/delphi/metrics/BasicMetricTest.java
@@ -74,15 +74,15 @@ public class BasicMetricTest {
 
   @Test
   public void testComments() {
-    assertThat((int) source.getMeasure(Metric.COMMENT_LINES), is(14));
-    assertThat((int) source.getMeasure(Metric.COMMENT_BLANK_LINES), is(1));
-    assertThat((int) source.getMeasure(Metric.PUBLIC_DOC_API), is(2));
+    assertThat(source.getMeasure(Metric.COMMENT_LINES), is(14));
+    assertThat(source.getMeasure(Metric.COMMENT_BLANK_LINES), is(1));
+    assertThat(source.getMeasure(Metric.PUBLIC_DOC_API), is(2));
   }
 
   @Test
   public void testLines() {
-    assertThat((int) source.getMeasure(Metric.LINES), is(72));
-    assertThat((int) source.getMeasure(Metric.LINES_OF_CODE), is(43));
+    assertThat(source.getMeasure(Metric.LINES), is(72));
+    assertThat(source.getMeasure(Metric.LINES_OF_CODE), is(43));
   }
 
 }
